{Add} Création de ConstructionStats.json
Permet d'automatiquement mettre les valeurs des constructions à jour suivant leur niveau et rareté.

Le fichier json possède des valeurs de test qui sont aléatoires.

Par simplicité, toutes les valeurs sont des floats (donc modification des variable de dégâts et de vieMax qui étaient des int)

Importation du Package Newtonsoft.Json pour facilement récupérer les données dans le json.

+++ Assets/Scripts/Enemy/Enemy.cs
@@ -6,7 +6,7 @@ public class Enemy : MonoBehaviour, ICanTakeDamage {
     public event EnemyDestroyedHandler OnDestroyed;
     
     [Header("Statistiques de base")] public int Level = 1;
-    [SerializeField] public int health = 200;
+    [SerializeField] public float health = 200;
     [SerializeField] public float moveSpeed = 1f; // Vitesse de déplacement de l'ennemi
     [SerializeField] public float timeBeforeBeingCible = 2f;
     [SerializeField] public int damageEnemy = 50;
@@ -113,7 +113,7 @@ public class Enemy : MonoBehaviour, ICanTakeDamage {
         OnDestroyed?.Invoke(gameObject);
     }
     
-    public void TakeDamage(int damage) {
+    public void TakeDamage(float damage) {
         health = Mathf.Max(0, health - damage);
 
         if (health <= 0) {
